requirements.txt
requests
python-dotenv
googlesearch-python
google-generativeai
beautifulsoup4
supabase
